refactor(ssg): move the outbound marker into its own module
reader_chrome.rs is at the repository's 350-line cap, so the marker
constant and the block-content check live beside it rather than in it.
